+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed the link name in the channel record page.
 - Fixed a typo in the `/channels` command.
 - Make the awards failure message translatable.
+- Apply fixes to translations.
+- Adds French translations.
+- Adds Spanish translations.
+- Adds Greek translations.
+- Adds German translations.
+- Adds some Tagalog translations.
 
 ## [v19.2.0](https://github.com/lexicalunit/spellbot/releases/tag/v19.2.0) - 2026-05-24