+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v19.3.0](https://github.com/lexicalunit/spellbot/releases/tag/v19.3.0) - 2026-05-25
+
 ### Fixed
 
 - Fixed the link name in the channel record page.